# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminSession {
  username: string;
  loggedIn: boolean;
}

export type TrackingEventStatus = typeof TrackingEventStatus[keyof typeof TrackingEventStatus];


export const TrackingEventStatus = {
  packed: 'packed',
  shipped: 'shipped',
  dispatched: 'dispatched',
  out_for_delivery: 'out_for_delivery',
  held_by_customs: 'held_by_customs',
  delivered: 'delivered',
} as const;

export interface TrackingEvent {
  id: number;
  shipmentId: number;
  status: TrackingEventStatus;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  note?: string | null;
  timestamp: string;
  createdAt: string;
}

export interface Shipment {
  id: number;
  trackingNumber: string;
  recipientName: string;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ShipmentWithEvents = Shipment & {
  events: TrackingEvent[];
};

export type TrackingResult = Shipment & {
  events: TrackingEvent[];
};

export interface ShipmentInput {
  /** @minLength 1 */
  recipientName: string;
  origin?: string;
  destination?: string;
  description?: string;
}

export interface ShipmentUpdate {
  /** @minLength 1 */
  recipientName?: string;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface TrackingEventInput {
  status: TrackingEventStatus;
  location?: string;
  note?: string;
  timestamp: string;
}

export interface TrackingEventUpdate {
  status?: TrackingEventStatus;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  note?: string | null;
  timestamp?: string;
}

export interface QuoteRequestInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  message: string;
}

export interface QuoteRequestUpdate {
  isRead: boolean;
}

export interface QuoteRequest {
  id: number;
  name: string;
  email: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}