# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteRequest {
  id: number;
  name: string;
  email: string;
  message: string;
  isRead: boolean;
  createdAt: Date;
}