# lib/api-zod/src/generated/types/shipment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Shipment {
  id: number;
  trackingNumber: string;
  recipientName: string;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
  updatedAt: Date;
}