# lib/api-zod/src/generated/types/shipmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentInput {
  /** @minLength 1 */
  recipientName: string;
  origin?: string;
  destination?: string;
  description?: string;
}