# lib/api-zod/src/generated/types/trackingEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackingEventStatus } from './trackingEventStatus';

export interface TrackingEvent {
  id: number;
  shipmentId: number;
  status: TrackingEventStatus;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  note?: string | null;
  timestamp: Date;
  createdAt: Date;
}