# lib/api-zod/src/generated/types/trackingEventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackingEventStatus } from './trackingEventStatus';

export interface TrackingEventInput {
  status: TrackingEventStatus;
  location?: string;
  note?: string;
  timestamp: Date;
}