# lib/api-zod/src/generated/types/trackingEventStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */

export type TrackingEventStatus = typeof TrackingEventStatus[keyof typeof TrackingEventStatus];


export const TrackingEventStatus = {
  packed: 'packed',
  shipped: 'shipped',
  dispatched: 'dispatched',
  out_for_delivery: 'out_for_delivery',
  held_by_customs: 'held_by_customs',
  delivered: 'delivered',
} as const;