# lib/api-zod/src/generated/types/trackingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shipvora courier tracking API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Shipment } from './shipment';
import type { TrackingEvent } from './trackingEvent';

export type TrackingResult = Shipment & {
  events: TrackingEvent[];
};